Backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
dotenv==0.9.9
eansearch==1.8.3
ecdsa==0.19.1
fastapi==0.119.1
Flask==3.1.0
flask-cors==5.0.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
postgrest==2.22.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
realtime==2.22.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
